refactor tela adm orcm
git-svn-id: https://localhost/svn/bruno_projects/trunk@30 d58b91e0-45a2-6f42-9374-62d56d83aa9a

+++ cashbuilder/grails-app/domain/com/cashbuilder/OrcmItem.groovy
@@ -4,7 +4,7 @@ class OrcmItem {
  
 	Categoria categoria
 	Subcategoria subcategoria
-	int valorPrevisto
+	double valorPrevisto
 	
     static constraints = {
     }
